backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfminer.six==20231228
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.0.0
transformers==4.40.2
torch==2.6.0
faster-whisper==1.0.1
